Handy implementations that come built-in

Instances

Justitia.AbstractMOIBackedInstance — Type

Instance that is internally stored as an MathOptInterface model.

abstract type AbstractMOIBackedInstance <: AbstractInstance end

To interact with the underlying MathOptInterface model, use:

get_moi_model(model::AbstractMOIBackedInstance)::MOI.ModelLike

Justitia.SimpleMOIBackedInstance — Type

A "simple" AbstractMOIBackedInstance where the underlying MathOptInterface model is stored in a field.

struct SimpleMOIBackedInstance{T <: MOI.ModelLike} <: AbstractInstance
    model::T
end

Justitia.MIPLIBInstance — Type

A MIPLIB 2017 instance stored in a MathOptInterface backing model.

MIPLIBInstance(instance_name::String)

Pulls down the MIPLIB instance named instance_name from the official MIPLIB website using curl.

Some pieces of the functional interface are implemented:

prep_instance!(instance::MIPLIBInstance)

Note

You must call prep_instance! on a MIPLIBInstance to populate the internal MathOptInterface model backing it.

Approaches

Justitia.MOIBasedApproach — Type

An approach that, given an instance, will merely convert that instance into a MathOptInterface.ModelLike. For instance, this could encapsulate a AbstractOptimizer, which ingests an instance and passes that instance to the underlying solver through the MathOptInterface API.

MOIBasedApproach{T}() where {T <: MOI.AbstractOptimizer}
MOIBasedApproach{T}(factory::Function) where {T <: MOI.AbstractOptimizer}

The underlying MathOptInterface will either be inferred from T, or you can optionally pass a factory function to the constructor. This factory must respect the signature

(instance::AbstractInstance, config::Dict{String,Any}) --> MathOptInterface.AbstractOptimizer.

Some pieces of the functional interface are implemented:

function build_model(
    approach::MOIBasedApproach{T},
    inst::AbstractMOIBackedInstance,
    config::Dict{String,Any} = Dict{String,Any}(),
) where {T <: MathOptInterface.AbstractOptimizer}

Models

Justitia.MOIModel — Type

A "simple" model that merely wraps around an underlying MathOptInterface model.

struct MOIModel{T <: MOI.AbstractOptimizer} <: AbstractModel
    opt::T
end

Some pieces of the functional interface are implemented:

optimize!(model::MOIModel)
tear_down(model::MOIModel, ::Type{MILPResult})

Results

Justitia.MILPResult — Type

Records solution information from a mixed-integer linear programming (MILP) problem.

MILP(;
    termination_status::MOI.TerminationStatusCode,
    primal_status::MOI.ResultStatusCode,
    x::Union{Nothing,Vector{Float64}} = nothing,
    primal_bound::Float64,
    dual_bound::Float64,
    solve_time_sec::Float64;
    node_count::Int = -1,
    simplex_iters::Int = -1
)

Records why the solver terminated (termination_status), if it had a feasible primal solution (primal_status), what that solution was if available (x), what the best primal cost was (primal_bound), the best available dual bound (dual_bound), elapsed solve time in seconds (solve_time_sec), and optionally the number of enumerated nodes (node_count) and total simplex iterations (simplex_iters).

Some pieces of the functional interface are implemented:

record_result(table::CSVRecord, result::MILPResult)

Additionally, there is a helper function for creating a CSVRecord with the correct headers:

CSVRecord(filename::String, ::Type{MILPResult})

Result Tables

Justitia.CSVRecord — Type

A simple AbstractResultTable that uses a .csv file as a backend.

CSVRecord(filename::String, headers::Union{Nothing,Vector{String}}=nothing)

Creates a CSVRecord table which writes the results to a .csv file at the location filename in the file system. Optionally, pass headers specifying the names of the fields to be recorded and the first line of the created .csv will include this header information.